requirements.txt
apturl==0.5.2
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
build==1.2.2.post1
cachecontrol==0.14.2
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.4.2
chrome-gnome-shell==0.0.0
cleo==2.1.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
contourpy==1.1.1
crashtest==0.4.1ls
cryptography==2.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.9
distro==1.4.0
distro-info==0.23+ubuntu1.1
dulwich==0.21.7
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
fastjsonschema==2.21.1
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
future==0.18.2
httplib2==0.14.0
idna==2.8
importlib-metadata==8.5.0
importlib-resources==6.4.5
installer==0.7.0
jaraco.classes==3.4.0
jeepney==0.9.0
jinja2==3.1.6
joblib==1.4.2
keyring==24.3.1
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lightning-utilities==0.11.9
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.5
matplotlib==3.7.5
menulibre==2.2.1
monotonic==1.5
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
netifaces==0.10.4
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
olefile==0.46
packaging==25.0
pandas==2.0.3
paramiko==2.6.0
pbr==6.1.1
pexpect==4.9.0
Pillow==7.0.0
pkginfo==1.12.1.2
platformdirs==4.3.6
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
protobuf==3.6.1
psutil==5.5.1
ptyprocess==0.7.0
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.1.4
pyproject-hooks==1.2.0
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
pytz==2025.2
pyxdg==0.26
PyYAML==5.3.1
rapidfuzz==3.9.7
reportlab==3.5.34
requests==2.32.3
requests-toolbelt==1.0.0
requests-unixsocket==0.2.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
SecretStorage==3.3.3
shellingham==1.5.4
simplejson==3.16.0
six==1.14.0
sympy==1.13.3
system-service==0.3
systemd-python==234
testresources==2.0.2
threadpoolctl==3.5.0
tomli==2.2.1
tomlkit==0.13.2
torch==2.4.1
torchmetrics==1.5.2
torchvision==0.19.1
triton==3.0.0
trove-classifiers==2025.5.1.12
typing-extensions==4.13.2
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.30.0
wadllib==1.3.3
xkit==0.0.0
zipp==3.20.2
